Deleted Player State Machine
- Deleted Player State Machine and put all functionality in PlayerController.cs to facilitate the networking testing and implementation.
