--- Assets/Development/Scripts/Player/States/Player/PlayerBaseState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 71de8dc8c5884dd4ea7a9283712f3654
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
