+++ Assets/Development/Scripts/Player/States/Player/PlayerController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 5ade8f8084c7edd47b4756da842f1362
+guid: 2278603552e903d4692cf7ad91790e32
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
