--- Assets/Development/Scripts/Player/States/Player/PlayerGroundState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: bfd9c84a5770f2446ba002bb0458e3de
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
