--- Assets/Development/Scripts/Player/States/Player/PlayerJumpingState.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 47436346a4729084ea5df0361444b433
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
