--- Assets/Development/Scripts/Player/States/Player/PlayerStateMachine.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: c22cb47df71950d4e943593d795de5b3
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
