--- Assets/Development/Scripts/Player/States/State.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 1fd425189b4b81d4c942fdd74a34d53d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
