--- Assets/Development/Scripts/Player/States/StateMachine.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 538083b768a6500409990a0d6467c2f1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
