+++ Assets/Development/Scripts/Stations/WaterStation.cs
@@ -27,7 +27,7 @@ public class WaterStation : BaseStation, IHasProgress
         slidertemp.value = 0;
     }
 
-    public override void Interact(PlayerStateMachine player)
+    public override void Interact(PlayerController player)
     {
         if (!HasIngredient())
         {
@@ -52,7 +52,7 @@ public class WaterStation : BaseStation, IHasProgress
         }
     }
 
-    public override void InteractAlt(PlayerStateMachine player)
+    public override void InteractAlt(PlayerController player)
     {
         if (ingredientListSOIndex >= ingredientListSO.Length - 1)
         {
