+++ Assets/Scenes/JuanTestScene.unity
@@ -21422,8 +21422,8 @@ MonoBehaviour:
   - {fileID: 1065496446}
   offset: {x: 0, y: 70, z: -70}
   smoothTime: 0.5
-  minZoom: 40
   maxZoom: 25
+  minZoom: 40
   zoomLimiter: 50
 --- !u!1 &1169943585
 GameObject:
@@ -39298,10 +39298,6 @@ PrefabInstance:
       propertyPath: m_Mesh
       value: 
       objectReference: {fileID: 530255702}
-    - target: {fileID: 3849241181019753718, guid: c431519c3e5173d4fac4acc61d044123, type: 3}
-      propertyPath: m_VersionIndex
-      value: 1881
-      objectReference: {fileID: 0}
     - target: {fileID: 3923269635416003139, guid: c431519c3e5173d4fac4acc61d044123, type: 3}
       propertyPath: m_Mesh
       value: 
@@ -39310,10 +39306,6 @@ PrefabInstance:
       propertyPath: m_Mesh
       value: 
       objectReference: {fileID: 1963772525}
-    - target: {fileID: 8606739759495854924, guid: c431519c3e5173d4fac4acc61d044123, type: 3}
-      propertyPath: m_VersionIndex
-      value: 2749
-      objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: c431519c3e5173d4fac4acc61d044123, type: 3}
 --- !u!65 &1079507868814769303
